+++ modules/alloy/src/main/ts/ephox/alloy/ui/types/SplitToolbarBaseTypes.ts
@@ -19,7 +19,7 @@ export interface SplitToolbarBaseApis {
   setGroups: (toolbar: AlloyComponent, groups: SketchSpec[]) => void;
   refresh: (toolbar: AlloyComponent) => void;
   toggle: (toolbar: AlloyComponent) => void;
-  isOn: (toolbar: AlloyComponent) => boolean;
+  isOpen: (toolbar: AlloyComponent) => boolean;
 }
 
 export interface SplitToolbarBaseSpec extends CompositeSketchSpec {