+++ modules/tinymce/src/themes/silver/main/ts/Render.ts
@@ -317,12 +317,12 @@ const setup = (editor: Editor): RenderInfo => {
     OuterContainer.focusToolbar(outerContainer);
   });
 
-  editor.addCommand('ToggleToolbar', () => {
-    OuterContainer.toggleToolbar(outerContainer);
+  editor.addCommand('ToggleToolbarDrawer', () => {
+    OuterContainer.toggleToolbarDrawer(outerContainer);
     // TODO: Consider firing event - TINY-6371
   });
 
-  editor.addQueryStateHandler('ToggleToolbar', () => OuterContainer.isToolbarToggled(outerContainer));
+  editor.addQueryStateHandler('ToggleToolbarDrawer', () => OuterContainer.isToolbarDrawerToggled(outerContainer));
 
   const mothership = Gui.takeover(
     outerContainer